Refactored classes to reflect correct Challenge
