+++ challenge-3/src/main/java/com/siemens/copilothackathonchallenge3/CoPilotHackaThonChallenge3Application.java
@@ -1,13 +1,13 @@
-package com.siemens.copilothackathonchallenge4;
+package com.siemens.copilothackathonchallenge3;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class CoPilotHackaThonChallenge4Application {
+public class CoPilotHackaThonChallenge3Application {
 
     public static void main(String[] args) {
-        SpringApplication.run(CoPilotHackaThonChallenge4Application.class, args);
+        SpringApplication.run(CoPilotHackaThonChallenge3Application.class, args);
     }
 
 }
