+++ challenge-3/src/main/java/com/siemens/copilothackathonchallenge3/controller/ProductUIController.java
@@ -1,7 +1,7 @@
-package com.siemens.copilothackathonchallenge4.controller;
+package com.siemens.copilothackathonchallenge3.controller;
 
-import com.siemens.copilothackathonchallenge4.entity.Product;
-import com.siemens.copilothackathonchallenge4.service.ProductService;
+import com.siemens.copilothackathonchallenge3.entity.Product;
+import com.siemens.copilothackathonchallenge3.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
